Add initial GHA setup
This patch brings simple initial pipelines in the realm of GitHub
Actions CI/CD workflows. It is not perfect but better than nothing.

Currently, a part of the tests is failing under Windows but all of
them pass under GNU/Linux (Ubuntu) and macOS.

Among other caveats: the deps are installed with `npm i` because the
lockfile is out-of-sync. This needs to be addressed separately and
replaced by an invocation of `npm ci`.

And also, another imperfection is that the testing is performed
against a non-packaged source. It is best to run tests against an
installed copy of the software to make sure that it is packaged
properly and contains all the necessary files (some may be excluded
from the tarball), ensuring that the users will get what's tested.

Things to explore in the future:
* https://stackoverflow.com/a/50222427/595220
* https://github.com/zkochan/package-preview
* Combining the use of manual unpacking of the tarball, copying tests
  into place and producing `npm shrinkwrap`

I've spent countless hours trying to making this work but had to stop
now. I hope that one day, it'll be possible to make this work properly.