chore: enforce npm and node requirements
Require modern node>=12 and npm>=7.11.2 in order to avoid
bugs and different behaviors during build and testing.

npm 7.11.2 is the version currently shipping with fedora 34.

That is in sync with requirements that we already use for
vscode-ansible extension. These requirements are build/test specific
and do not affect the runtime.

Prevents bug where older npm would mess the format of package-lock.json file.

Originally included in #17