Make runCommand strip ANSI codes
This makes runCommand automatically strip ANSI escapes from stdout
and stderr, while still allowing user to opt-in for not stripping
them.

The reason for using stripping by default is that the primary use
case for runCommand is to process the output and not to display it
inside a terminal. This should prevent bugs where we fail to parse
the output from various tools due to presence of ANSI codes.

Related: https://github.com/ansible/vscode-ansible/issues/373