Reformat typescript using prettier
That enables typescript on prettier as currently almost any attempt
to edit .ts files using vscode would introduce undesired reformatting.
That is because we recommend prettier extension which auto-reformats
but without checking that in CI.

The only manually made change is the one line in pre-commit config,
the rest is just pre-commit itself.