Standardize documentation building
Use a single job and tox target for building documentation, just as we
have with other devtools projects. Building documentation should be
do with a single command, not a chain of them.

This also declutters the CI/CD pipelines, thus improving the the
performace and reduce the resource waste.

Related: https://github.com/ansible-community/devtools/issues/5